+++ CHANGELOG.md
@@ -17,8 +17,11 @@ Types of changes
 ## [2.6.0]
 
 - `Added` `sqlserver` support
-- `Added` flags `sample-size`, `distinct`, `limit`, `where` and `table` to the `analyse` command, by default distinct values are not counted
-- `Added` flags `exlude` and `exlude-pk` to exclude columns from analyse
+- `Added` flags to `lino analyse` : `sample-size`, `distinct`, `limit`, `where` and `table` (by default distinct values are not counted)
+- `Added` flags to `lino analyse` : `exlude` and `exlude-pk` to exclude columns from analyse
+- `Added` flags to `lino pull` : `-X` or `--exclude-from-file` to filter out lines from a JSON lines file reference
+- `Added` flags to `lino push` : `--savepoint` to write primary keys of all commited line to a file
+- `Added` new import types to `table.yaml` : `blob` will optimize push of blob column into database, `file` will read content of a file at the specified path and import as a blob
 
 ## [2.5.0]